pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["ajib6ept <asAlbert@narod.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.1"
dj-database-url = "^0.5.0"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
